Split GameBox.cs into that and GameBoxOfT.cs
